refactor: extract governance enums
