+++ packages/nns/src/utils/neurons.utils.spec.ts
@@ -1,9 +1,10 @@
-import { NeuronInfo, ProposalInfo, Vote } from "../types/governance_converters";
+import { NeuronInfo, ProposalInfo} from "../types/governance_converters";
 import {
   ineligibleNeurons,
   votableNeurons,
   votedNeurons,
 } from "./neurons.utils";
+import {Vote} from '../enums/governance.enums';
 
 describe("neurons-utils", () => {
   const proposalTimestampSeconds = BigInt(new Date().getTime());
